Extract ActionButton from DAE
Use button factory on perspectives:
- Beam Status
- DAE, Run Summary Tab
