+++ base/uk.ac.stfc.isis.ibex.ui.beamstatus/META-INF/MANIFEST.MF
@@ -22,7 +22,8 @@ Require-Bundle: org.eclipse.ui,
  org.csstudio.trends.databrowser2;bundle-version="4.2.4",
  uk.ac.stfc.isis.ibex.logger,
  uk.ac.stfc.isis.ibex.ui.configserver,
- javax.annotation;bundle-version="1.3.5"
+ javax.annotation;bundle-version="1.3.5",
+ uk.ac.stfc.isis.ibex.ui.widgets;bundle-version="1.0.0"
 Bundle-RequiredExecutionEnvironment: JavaSE-17
 Export-Package: uk.ac.stfc.isis.ibex.ui.beamstatus.views
 Import-Package: org.eclipse.wb.swt,
