+++ base/uk.ac.stfc.isis.ibex.ui.dae/src/uk/ac/stfc/isis/ibex/ui/dae/run/RunSummary.java
@@ -41,6 +41,7 @@ import uk.ac.stfc.isis.ibex.logger.IsisLog;
 import uk.ac.stfc.isis.ibex.ui.dae.DaeUI;
 import uk.ac.stfc.isis.ibex.ui.dae.experimentsetup.ExperimentSetupViewModel;
 import uk.ac.stfc.isis.ibex.ui.dae.widgets.LogMessageBox;
+import uk.ac.stfc.isis.ibex.ui.widgets.buttons.IBEXButtonFactory;
 import uk.ac.stfc.isis.ibex.ui.widgets.observable.WritableObservingTextBox;
 
 /**
@@ -173,17 +174,12 @@ public class RunSummary {
 
 		new Label(infoComposite, SWT.NONE);
 		new Label(infoComposite, SWT.NONE);
-
-		btnDisplayTitle = new Button(infoComposite, SWT.CHECK);
-		btnDisplayTitle.setLayoutData(new GridData(SWT.BEGINNING, SWT.CENTER, false, false));
-		btnDisplayTitle.setText("Show Title and Users in Dataweb Dashboard Page");
-		btnDisplayTitle.addSelectionListener(new SelectionAdapter() {
-			@Override
-			public void widgetSelected(SelectionEvent e) {
-				super.widgetSelected(e);
-				model.displayTitle().uncheckedSetValue(btnDisplayTitle.getSelection());
-			}
-		});
+		
+		btnDisplayTitle = IBEXButtonFactory.checkbox(infoComposite, "Show Title and Users in Dataweb Dashboard Page", 
+				"Show Title and Users in Dataweb Dashboard Page", 
+				evt -> {
+					model.displayTitle().uncheckedSetValue(btnDisplayTitle.getSelection());
+				});
 
 		messageBox = new LogMessageBox(lhsComposite, SWT.NONE);
 		messageBox.setLayoutData(new GridData(SWT.FILL, SWT.FILL, true, true, 1, 1));
