+++ base/uk.ac.stfc.isis.ibex.ui.widgets/META-INF/MANIFEST.MF
@@ -14,5 +14,6 @@ Require-Bundle: org.eclipse.ui;bundle-version="3.7.0",
  org.eclipse.jface.databinding;bundle-version="1.5.0",
  uk.ac.stfc.isis.ibex.model;bundle-version="1.0.0",
  uk.ac.stfc.isis.ibex.logger
-Export-Package: uk.ac.stfc.isis.ibex.ui.widgets
+Export-Package: uk.ac.stfc.isis.ibex.ui.widgets,
+ uk.ac.stfc.isis.ibex.ui.widgets.buttons
 Automatic-Module-Name: uk.ac.stfc.isis.ibex.ui.widgets
